chore: 添加格式忽略与 mock 数据
