+++ .prettierignore
@@ -1,0 +1,1 @@
+docs/工作流编排_studio（模拟页面）.jsx
